extend-rabitq/web/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
markdown-it-py>=3.0
